+++ component/helpers.cmake
@@ -34,3 +34,6 @@ include("${CMAKE_CURRENT_LIST_DIR}/materialize.cmake")
 include("${CMAKE_CURRENT_LIST_DIR}/backend/helpers.cmake")
 
 include("${CMAKE_CURRENT_LIST_DIR}/factory.cmake")
+
+# Verbose configure dump (after the graph API exists).
+include("${CMAKE_CURRENT_LIST_DIR}/../report/helpers.cmake")